mnt/user-data/outputs/nlp-project/src/requirements.txt
nltk==3.8.1
textblob==0.17.1
pytest==7.4.3
pytest-cov==4.1.0
pandas==2.1.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
